chore: move powerset to diarization task
